Enhance Sprot buffer wrappers
The message construction and parsing built into the `TxMsg` and `RxMsg`
types and their corresponding `VerifiedTx` and `VerifiedRx` versions,
was not quite completed when merged into #965. Both @jgallagher
and @mkeeter requested further changes to make them more robust on that
PR, but the changes were merged in an intermediate state. See the
following comments for details:

 * https://github.com/oxidecomputer/hubris/pull/965#discussion_r1035271064
 * https://github.com/oxidecomputer/hubris/pull/965#discussion_r1035225087
 * https://github.com/oxidecomputer/hubris/pull/965#discussion_r1035257378
 * https://github.com/oxidecomputer/hubris/pull/965#discussion_r1035280375
 * https://github.com/oxidecomputer/hubris/pull/965#discussion_r1035283082

This commit changes the message types to borrow the underlying buffers
mutably in each message type so that only one code path at a time can
use the buffers. The TxMsg serialization functions also consume self
and return a VerifiedTxMsg to prevent mutation of the underlying buffer.
The RxMsg type ties header validation to checksum verification in the
consuming `parse` method taht returns a VerifiedRxMsg. These changes
alone satisfy @mkeeter's concerns.

The types also track the amount of buffer used satisfying one of
@jgallagher's requests, and removing the need to track that in a
variable outside the types.

More construction methods were added to `TxMsg`, but not all
mutation of the payload was removed as that seems impractical. So the
last of @jgallagher's requests here were partially satisfied.

This commit also utilizes the changes to `TxMsg`, `RxMsg`,
`VerifiedTxMsg`, and `VerifiedRxMsg` from the prior commit in the
stm32h7-sprot-server.

Additionally:

 * RAII is now used with the SPI API, such that manual release
of the lock is no longer needed.
 * Request sending is simplified and wrapped in a function
 * IRQ related code was broken out into functions and comments moved to
function headers
 * Duplicate checks for error conditions in do_send_recv_retries
were removed from rot_sink, simplifying it.
 * Use of cfg_if was removed for the sink test code so that we could
   de-indent it, and use rustfmt.
 * Retry timeouts were added before retrying after a recoverable error.

There was also an extra constraint put on the rot_sink test: It now only
supports buffers of two bytes or larger. This was just for simplicity,
and it can be reverted if necessary.

This commit also utilizes the changes to `TxMsg`, `RxMsg`,
`VerifiedTxMsg`, and `VerifiedRxMsg`, in the lpc55-sprot-server.

Additionally:
 * The RoT handler code no longer owns IoStats and some new stats have
been added
 * StartupState is extracted into its own type
 * The RoT handler is simplified, and duplicate matching on unexpected
messages has been removed.
 * Interrupt management has been simplified such that the only interrupt
waited for by the lpc55-sprot-server is `ssa` or CSn asserted.
 * Support for reading the `ssd` flag out of the stat register was added
to support using only the `ssa` interrupt.
 * Various functionality has been split into its own functions.