Embed `SpiServerCore` into tasks with exclusive peripheral ownership (#1086)
This PR allows for tasks to exclusively own their SPI peripheral, instead of
talking to a separate SPI server task.

It includes code changes to the following 4 tasks:
- `drv-stm32h7-sprot-server`
-  `task-monorail-server`
-  `drv-fpga-server`
-  `task-net`

There exist other tasks which use SPI but don't have exclusive use of a
peripheral, e.g. `gimlet-seq-server`; such tasks are left (mostly) unchanged for
now, except where needed to support the new parameter in `SpiDevice`

In general, owning the SPI peripheral is enabled with the `use-spi-core`
feature, which also requires wiring up the SPI IRQ, adding it to the `uses =
[...]` list, and adding `config.spi.global_config = "spiX"` to the task. This is
checked at compile time!

In addition to the code changes, this PR updates a bunch of `app.toml` files to
eliminate separate SPI servers whenever possible.