Plumb SPI features through tasks to improve rebuild (#1104)
Previously, Sidecar took 19 seconds to rebuild, because
`drv-stm32h7-spi-server-core` was being

- instantiated more than once
- with identical features
- but dependent on environmental variables that are changing

This caused it to be rebuilt – and more importantly, the tasks using it to be
re-linked – multiple times per full Hubris build.

Because Cargo caches builds based on features, we can fix this by activating a
`spiX` feature in this crate, based on the SPI peripheral being used.

This requires a little plumbing to propagate the feature through other crates
which depend on `drv-stm32h7-spi-server-core`.

This brings a Sidecar rebuild down to **7 seconds**, which is much more
tolerable.

In addition, the feature is used to select the global SPI configuration block,
eliminating the old `global_config`.  This couples the global configuration
block name to SPI peripherals, which seems... fine?  We were already naming
those blocks based on peripherals, and this removes a bunch of copy-pasta.