HACK: bit-bang kernel trace events on gimlet LED
There's limited easy-to-probe GPIO from the SP on gimlet B-rev and
later, much of which is only on the underside of the board. So, to get
trace information out, we can't do a parallel trace to test points like
we did before.

Instead, this bit-bangs asynchronous serial on the indicator LED. Sample
this at the high side of the pulldown on the LED FET.

Decode this at 33 125 000 baud, 16 bits per frame, 1-2 stop bits.