hash: remove some sleeps
The hash driver resets the peripheral for every hash, and sleeps for 1ms
between setting and resetting the reset line. Since hashes are
relatively fast, this was adding a nontrivial amount of delay to every
hash operation. Delays around resetting peripherals are not necessary on
STM32, so I've removed them.