stm32xx-i2c-server: don't glitch pins on mux swaps
By alternating between `configure_alternate` (for normal operation) and
`configure_input` (to shut off pins and switch to a different set of
pads), the I2C driver was introducing a subtle and brief glitch in the
output lines -- as a side effect of how `configure_input` is defined.

Concretely, `configure_input` implies setting the pin's alternate
function mux (AF) to an arbitrary value, under the assumption that the
configuration will be set once and not changed.

The `configure_alternate_input` operation lets us preserve the AF during
reconfiguration and should prevent the glitch. However, the STM32 docs
are ambiguous on whether the peripheral selected by the AF mux can still
see inputs when the pin is set to input -- and the block diagrams
strongly suggest that it can.

And so, rather than input, this change sets it to _analog_ mode. This
disconnects the input buffer completely, preventing _anything_ in the
chip from responding to the state. We can do this while preserving the
AF mux to avoid the glitch when switching back.

This glitch appears to date to the original I2C commit, but it seems we
haven't noticed it until now -- all the more reason to have people
looking at behavior on scopes!