arm_m: clear SVC/fault pend bits if task faults.
This fixes an unusual but vexing misbehavior in fault handling. If a
task ran out of stack as a side effect of executing an SVC instruction,
instead of entering the kernel at SVCall, it would enter at
MemManageFault. However, _the SVC exception triggered by the SVC
instruction is still pending when this happens._

This means that the flow of control goes like this:

1. Task faults.
2. Kernel does bookkeeping.
3. Kernel notifies supervisor and switches to supervisor task.
4. Kernel attempts to return to taskyland.
5. ARM M-profile tail chaining kicks in, converting the
   return-from-interrupt operation into a branch to SVCall.
6. Because we've already changed the current task, this simulates a
   surprise syscall _from the supervisor task._
7. In practice, the syscall is recv, and the supervisor then issues a
   bogus-shaped recv instead of handling the fault.

We can fix this by clearing the SVCALLPENDED bit in the interrupt
controller. In practice, we can do this unconditionally when we're in
the kernel handling any task fault -- since a fault in the kernel while
servicing an SVC would follow a different (and fatal) code path.

And so that's what I've done here.

We also do this to the other fault pending bits, since the failure
described for SVC above can also apply to any illegal instruction
generating a Usage fault.

Thanks to @mkeeter for working out the test reduction case for
reproducing this on any H7 board, and @lzrd and @kc8apf for helping me
think through the exception states.

Fixes #1134.