stm32h7-qspi: eliminate busy wait on completion
The QSPI peripheral provides an interrupt for Transfer Complete, but not
for Busy being cleared. I originally interpreted the QSPI docs as
requiring us to wait for Busy to be cleared before doing more things,
which left us with this odd duck of a polling loop. It used the Busy
flag as its loop exit condition, while attempting to sleep on the
Transfer Complete interrupt.

However, in practice, Transfer Complete gets set a ways before Busy is
cleared. Once TCF is set in SR, this loop fails to sleep, waking
immediately on every iteration and burning CPU.

A re-read of the docs (and testing on a Gimlet rev B under flash test)
suggests that the reason Busy doesn't have an interrupt is that we don't
actually need to wait for it -- commands are issued through a FIFO, so
there's no problem with us starting the next operation while Busy is
set, so long as Transfer Complete has happened.

So, this switches the loop to use TC as the exit condition as well as
the wait condition, eliminating the busy-loop.