Initial support for host flash persistence (#1145)
This PR implements the scheme described in
https://github.com/oxidecomputer/hubris/issues/985 for storing persistent host
flash data (current slot and boot flags).

It reserves the lowest sector (64 KiB) of QSPI flash for Hubris, adding new HIF
functions if you _really_ want to overwrite it and returning an error code
otherwise.  Within sector 0, we store up to 512 copies of `RawPersistentData`;
the one with the largest `monotonic_counter` field is our current value.

A new Idol API is added to write this persistent data, by scanning sector 0 on
both flash ICs, finding the highest `monotonic_counter`, and incrementing it.

`control-plane-agent` is modified to avoid writing sector 0, but checks that any
bytes that we _try_ to write in that sector are `0xFF` and returns an error
otherwise.