Add `Power.read_mode` (#1152)
Currently, we read the BMR491 black-box through raw I2C in Humility.

We'd like to switch to using the `bmr491_*` APIs, which are Idol functions that
could be called remotely – initially via `humility rpc`, and later by MGS.

However, to interpret IBC blackbox data, we need to know `VOUT_MODE`.  This PR
adds an Idol function to read it!