Add the Hubris Caboose (#1169)
Closes #1157, which has discussion of this design.

This PR does several things!

Primarily, it adds an optional `caboose` to Hubris images.  The caboose is
located at the end of flash and contains (1) two words indicating that it's a
caboose and (2) extra space for user data.

It is documented in a newly added section of the [Hubris
manual](https://github.com/oxidecomputer/hubris/blob/caboose/doc/guide/caboose.adoc).

This also accomplishes _some_ of the goals of #1004 , namely having a "commit
point" at the end of flashing.  For the moment, this implementation still
requires having an `ImageHeader` after the vector table.  The caboose is after
everything else in flash, but is not at the last page; this means that we need a
way to find it, which is specified in the `ImageHeader`.

(We could work around this by using NXP-style overloading spare vector table entries:
https://github.com/oxidecomputer/hubris/issues/1157#issuecomment-1440878674
but this ends up feeling hacky on the STM32 and tightly couples us to a
Cortex-M-flavored MCU)

--------------------------------------------------------------------------------

In addition, this PR moves common Hubris archive functionality to a new
[`hubtools` repository](https://github.com/oxidecomputer/hubtools).  This is
because both Hubris `xtask` and `permission-slip` want to do certain
manipulations of Hubris archives – most notably, going from `srec` to all the
other binary formats.

`hubtools` contains this common functionality, as well as a minimal CLI to edit
the caboose.

--------------------------------------------------------------------------------

Finally, the PR adds a new `caboose-reader` task, which is a minimal task to
read from the caboose.  It's mostly useful for development and debugging.