Make DICE SYSCON accesses explicit and volatile.
Previously, the dice crate was creating aliasing &mut[]s to the SYSCON
whenever you called certain operations. This was potentially unsound,
but fortunately it's also not strictly necessary: we can instead create
slices of the VolatileCell type used in the PACs to model machine
registers.

This also fixes a potential miscompile where we were representing the
registers as &mut [u32], which -- on ARMv8-M where unaligned accesses
are cheap -- permits the compiler to use any size accesses to that
memory, because the type says it's normal memory. VolatileCell fixes
that.

I've got callers passing in the SYSCON now so we can track the borrow
(and prove they have access in the first place!).

For this to work I had to manually implement ZeroizeOnDrop, which
highlighted the fact that it's trivial to break security properties by
removing a derive(ZeroizeOnDrop) -- no code will break!