Disable ITM output in on root of trust builds.
We aren't using this in practice, it's not clear that it still works,
and it provides a permissionless (available to all tasks) transmit data
channel in production that we haven't really thought about securing.

Instead, I've moved the ITM setup code into the only routine that was
actually using it: the non-RoT LPC55 test suite.