update: add switch_default_image
This works on LPC55 with the RFD374 stage0 to flip the persistent boot
preference bit by rewriting the CFPA. I've rebased it onto master in the
hopes of getting it merged sooner. It works today, just systems built
from master will not currently _respect_ the setting it changes.

On STM32H7 it just fails, a side effect of trying to share the update
API.