Gimlet rev D support (#1237)
Closes #1210

This is 90% plumbing, and 10% updating the I2C build system to support `nreset`
pins for the PCA954x mux.  We already had an `enable` pin wired up, but it's
only used on the LTC4306 (on the `gemini-bu` board), so I changed the name to
match the common case going forward.