kern: rework dump impl to reduce UB
Doing a kernel dump from within the kernel is inherently hairy. We wind
up needing to do a raw byte read of a Task struct, which we also need to
interact with as an actual Task.

The implementation was
- Creating (using core::slice::from_raw_parts) a byte slice that aliased
  the &mut [Task] task table
- Then using it and the &mut concurrently

which is UB due to aliasing violations.

The code also contained an easily accessible way of crashing the kernel:
send in a region such that base+size overflows. Triggering this reliably
required choosing a base value within the TCB address range, because the
other code path used USlice.

I've reworked the logic. The particular code that was causing the UB
issue has been replaced by

- Deriving a &mut Task to the specific task we want to access, and then
- Deriving a &[u8] from it using transmute and shadowing to prevent
  aliasing.

(This was also the second time in the kernel I've wanted the
index2_distinct function, so I've pulled it into an accessible common
location.)

I've eliminated the caller-controlled kernel panic by using USlice along
all code paths. This had the side effect of removing some runtime
overflow checks from the generated code, slightly reducing code size.
(However, the code is slightly bigger overall as a result of this
change.)

I've also added a crapload of comments to try and help future readers of
this subtle routine.