[sprot] Use hubpack based messages across the SPI interface
In order to simplify the implementation of sprot, we base all sprot
protocol messages around hubpack serialization. There is a fixed sized
`Header`, and a parameterized `Msg` that can contain either a `ReqBody`
or `Result<RspBody, SprotError>`. This removes a ton of hand-written
serialization code that was quite hard to reason about. Overall it
should  be easier to extend the protocol with new messages now.

This change also changes the `SprotError` from a  `CLike` error to a
`Complex` error in idol terminology. We no longer translate sub-errors
into a global list of integers, but compose them into a `SprotError`
enum.

In order to improve performance on the RoT side, we know longer bother
transmitting bytes when receiving messages or pulling bytes out  of the
receive fifo when sending messages.

This change is a WIP because:
 1. Humility cannot yet handle the `Complex` idol errors, preventing
sprot from being used directly by humility.
 2. The SprotError type has not been plumbed up through the control
plane agent yet. After discussion with @cbiffle and @jgallagher, it has
been proposed that we put the `SprotError` type directly into `gateway-
messages`. For this to work properly though we'd likely have to also put
any sub-messages, such as `UpdateError`, into gateway-messages as well.
Alternatively, we could use `From` impls to convert between error types.