Add network feature to `dump-agent` (#1278)
This is part of the https://github.com/oxidecomputer/hubris/issues/1217 saga,
which begins with https://github.com/oxidecomputer/humpty/pull/4/files and also
requires a Humility PR.

This Hubris PR adds an optional `net` feature to `dump-agent`.  When the feature
is enabled, `dump-agent` listens on a UDP socket for messages which encode dump
commands, using API types from `humpty` (see PR above).  These messages are a
thin layer above the existing hiffy API.

Most of the code changes are in the `udp` module; changes in `ServerImpl` are
mostly to make more functions callable from either hiffy or the `udp` module.