reset during I2C transaction can induce mayhem (#1281)
We have seen a series of vexing VPD problems on reset, with a variety
of underlying causes.  While we fixed several of these, the problem
remained (see #1196) -- and one clear cause was a reset during an I2C
transaction, resulting in an errant first transaction after the reset.
This fix engages in the time-honored tradition of wiggling SCL on
reset, and then inducing a STOP condition to effectively terminate any
zombie I2C transactions.