[sprot] Use hubpack based messages across the SPI interface (#1279)
* [sprot] Use hubpack based messages across the SPI interface

In order to simplify the implementation of sprot, we base all sprot
protocol messages around hubpack serialization. There is a fixed sized
`Header`, and a parameterized `Msg` that can contain either a `ReqBody`
or `Result<RspBody, SprotError>`. This removes a ton of hand-written
serialization code that was quite hard to reason about. Overall it
should  be easier to extend the protocol with new messages now.

This change also changes the `SprotError` from a  `CLike` error to a
`Complex` error in idol terminology. We no longer translate sub-errors
into a global list of integers, but compose them into a `SprotError`
enum.

In order to improve performance on the RoT side, we know longer bother
transmitting bytes when receiving messages or pulling bytes out of the
receive fifo when sending messages. We also now use 16-byte fifos which
seems to eliminate all SPI errors when `ping` is not running.

Co-authored-by: Bryan Cantrill <bryan@oxide.computer>
Co-authored-by: Ben Stoltz <ben.stoltz@oxidecomputer.com>