Improve QueueFull watchdog (#1311)
Using `can_send` is too optimistic; it will return `true` if there's room in the
buffer but a max-length packet won't actually fit.  This results in a storm of
`QueueFull` if

a) The host isn't replying
b) There's room in the buffer, **but**
c) The next packet is larger than available space

This PR switches to the single source of truth for socket happiness: did we get
a `QueueFull`?