gimlet-hf: reduce clock speed to 33MHz.
We've been having flake with the QSPI flash on Gimlet C revs, some of
which has been manufacturing errors, but not all of it.

It turns out we've got at least one board (serial '43) that appears to
be assembled correctly, but whose QSPI circuit doesn't work at 40 MHz.
The bits wind up shifting with respect to the clock, suggesting that
we've got a propagation delay or rise time issue across the translator
and mux.

This change lowers the speed to 33 1/3 MHz, which is stable on all the
machines we've tried so far.