[sprot] Add API for rot_boot_info
This adds the sprot and update APIs for reading RoT boot preference
information out of CFPA. It carries over information about the currently
active RoT slot from the pre-main handoff data used in the status APIs.
Transient boot selection is not currently implemented, and so there is a
placeholder Option<SlotId> in the struct for that information.