Add per-component caboose reading (#1357)
This PR adds support for the `ReadComponentCaboose` message.

It plumbs the message from `control_plane_agent` all the way to the RoT's
`update_server`, which is actually responsible for reading values out of flash.

In passing, it splits the `idl/update.idol` into board-specific files, because
their differences were getting too large to bear; for example, the STM32H7
update server only reads the inactive bank caboose, while the RoT update server
reads both cabooses.

The RoT `Flash` peripheral has multiple functions change to take `&self` instead
of `&mut self`; this is because we must be able to clone the `TlvcRead`, which
in turn uses that peripheral to read data from flash.  This is not ideal, but
seems harmless enough in our single-threaded system.