Bump lpc55-update-server task size up by 512 (#1375)
Our stack margin prior to this PR was 24 bytes; after this PR, we usually see a stackmargin of 536 bytes (i.e., 24+512), but after running sufficiently-many requests for state from MGS, we see it drop to 432 bytes. Cliff suggests this is from getting an interrupt at an inopportune time. This explains what we were seeing in #1374: if we received such an interrupt while the update was running, it would trigger a SO of update_server, explaining the "no update started" error we saw.