Clear spurious CGBAD flags on the VSC8562 (#1406)
This flag is set when we first initialize the system and when we change link
speed, because the VSC7448 and VSC8562 are configured one after the other.  It's
spurious and does not reoccur, so it seems reasonable to clear it here.
Otherwise, we see a false error reported in `humility monorail status`, which
goes away when you re-run the command.