Add Jefe::restart_me and use it for net watchdog.
This adds a method for tasks to restart themselves without indicating a
fault, which has the advantage of bypassing Jefe's crash-dump logic for
deliberate restarts.

Currently the only case of a deliberate restart is in net, for the
receive watchdog. This should knock about 66ms off each firing of that
watchdog.

I initially implemented this as a syscall, but @mkeeter noted it could
be done entirely outside the kernel -- and so here it is. Long live the
emokernel.