transceivers-server: cleanup interface (#1446)
Most of the interface here was built out to facilitate benchtop testing during initial development. Now that we have these interfaces proven out and connected via the UDP interface, we don't really need them as part of the general Idol server interface. I'm opting to remove them completely as otherwise it just doubles the surface area where we need to manage the module interactions. This will leave the UDP interface as the sole way to manage the transceivers going forward.