attest: Add function to record a measurement observed by the platform.
This function is *not* idempotent: if the 'record' function is called
with the same parameters repeatedly the measurement log will eventually
fill up resulting in an error. The measurement log is just an array of
measurements (aka [u8; 32]) that is managed in an ArrayVec. The 'record'
function takes an enum parameter identifying the hash algorithm and a
lease holding the digest. We pass the digest as a lease partially because
humility / hiffy can't pass arbitrary arrays as parameters, but also
because it's possible that in the future we'll need to support digests
with varying lengths.