sp_measure: Remove unnecessary loop in main.
This task is intended to run once when the system boots and the task it
performs should not be repeated. Jefe will no longer restart this task
and so we could remove the final call to `sys_recv_closed` as well,
however tasks that return from their `main` function are reported as
having executed an illegal instruction. So we leave this call in place
to prevent a scary message from showing up in the task list: `FAULT:
illegal instruction (was: ready)`