sidecar: Fix comment about fan rotor ordering (#1445)
In a previous comment (https://github.com/oxidecomputer/hubris/pull/1400#discussion_r1223337891) I stated:

>The drawing we got from SD notes an inlet vs outlet fan. In our system the inlet fan is more "north" and connects to pwm0/tach0, while the outlet fan is more "south" and connects to pwm1/tach1.

This is incorrect, for whatever reason my brain that day was thinking that the inlet fan was more towards the back (aka North) while the outlet fan was towards the front (aka South). This is backwards as we draw air from front to back, not back to front, which means that the inlet fan is actually South while the outlet fan is North. This commit cleans up that error.

There aren't any other system impacts beyond fixing up the comment and fan naming.