lpc55-rot-startup: Use a better method to generate an FWID.
The current implementation captures all of the RoT software. It does not
however include other pages in the RoT image flash range. Expanding our
definition for the RoT FWID as not only the expected image, but also any
other programmed flash pages in the range of possible pages mitigates a
potential persistence mechanism for an attacker.

This commit:
- adds a new environment variable HUBRIS_FLASH_OUTPUT that holds all
  config::Output structures for the flash regions assigned to the image
  being built
- generates a Range for each flash region in HUBRIS_FLASH_OUTPUTS for
  use by the lpc55-rot-startup crate
- uses the A & B Ranges to walk through all possible flash pages for the
  given image including each page that's been programmed in the FWID