wip: lpc55-rot-startup: Use a better method to generate an FWID.
The current implementation captures all of the RoT software which is
good. It does not however include other pages in the RoT image flash
range. Expanding our definigion for the RoT FWID as not only the
expected image but also any other programmed flash pages in the
range of possible pages mitigates a potential persistence mechanism
for an attacker.

This implementation walks through all possible flash pages for imagea
(currently from hard coded (address, size)). All pages within this range
that have been programmed are included in the FWID. We do not parse the
data, we just roll it into the FWID hash calculation using the `update`
API.

TODO - Get flash size / end from build.