Sidecar: avoid Monorail restart if Tofino PCIe link not up (#1510)
Monorail monitors the 10G link with Tofino and periodically restarts the task if
this link is not up. This in turn causes the technician port PHY to flap which
is disruptive while working with `pilot racktest`.

In order to reduce spurious restarts the PCIe link with Tofino is monitored and
restarts are only performed if the link is up, assuming the 10G port is down
otherwise. In addition logging in the sequencer task is improved, making the
events more accurate and cutting down on noise.