Always calculate the firmware measurement even if there is no valid image.
Note: The measurement is not plumbed through to the management plane at
this time. It will wait for other API changes.

Also, figure out the programmed pages once while performing the slot measurement
and use those results when the image is being verified.

Tweak .gitignore for tags and cscope users.